no casting for data_frame in makeSyncFiles.py
Former-commit-id: fc71cba3c871fc33472f51030348e4856f25bc6a

+++ dannce/utils/makeSyncFiles.py
@@ -89,7 +89,7 @@ if fps > 1000:
 fp = 1000.0 / fps  # frame period in ms
 fp = fp.astype(int)
 
-data_frame = np.arange(framecount[0]).astype("float64")
+data_frame = np.arange(framecount[0])
 data_sampleID = data_frame * fp + 1
 data_2d = np.zeros((framecount[0], 2 * num_landmarks))
 data_3d = np.zeros((framecount[0], 3 * num_landmarks))
